Move Spree::StaticPage and Spree:StaticRoot to facilitate overrides with decorator
